Extract generic constraint classifier
Move shared generic-constraint presence and value-type eligibility policy out of LibraryBodyAnalysisBuilder while preserving reader-relative external metadata handling.
