+++ src/ILInspector.Analysis/LibraryBodyGenericConstraintClassifier.cs
@@ -1,0 +1,109 @@
+using System.Reflection;
+using System.Reflection.Metadata;
+using System.Reflection.Metadata.Ecma335;
+
+using ILInspector.Metadata;
+
+namespace ILInspector.Analysis;
+
+/// <summary>
+/// Classifies generic-constraint facts for primary-image optimization analysis
+/// and reader-relative async-sibling analysis.
+/// </summary>
+internal sealed class LibraryBodyGenericConstraintClassifier(
+    MetadataReader reader)
+{
+    readonly MetadataReader _reader = reader;
+
+    internal bool HasGenericConstraints(
+        MetadataReader declaringReader,
+        MethodDefinition method)
+    {
+        foreach (var handle in method.GetGenericParameters())
+        {
+            var parameter =
+                declaringReader.GetGenericParameter(handle);
+            if (parameter.Attributes
+                    != GenericParameterAttributes.None
+                || parameter.GetConstraints().Count > 0)
+            {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    internal bool GenericParameterCanBeValueType(
+        TypeRef genericParameter,
+        MethodIdentity caller)
+    {
+        try
+        {
+            var methodHandle = (MethodDefinitionHandle)
+                MetadataTokens.EntityHandle(caller.MetadataToken);
+            var method = _reader.GetMethodDefinition(methodHandle);
+            GenericParameterHandleCollection handles =
+                genericParameter.Kind == TypeRefKind.MethodGenericParameter
+                    ? method.GetGenericParameters()
+                    : _reader.GetTypeDefinition(method.GetDeclaringType())
+                        .GetGenericParameters();
+            if (genericParameter.GenericParameterIndex < 0
+                || genericParameter.GenericParameterIndex >= handles.Count)
+            {
+                return false;
+            }
+
+            var handle = handles.ElementAt(
+                genericParameter.GenericParameterIndex);
+            var parameter = _reader.GetGenericParameter(handle);
+            if ((parameter.Attributes
+                    & GenericParameterAttributes.ReferenceTypeConstraint) != 0)
+            {
+                return false;
+            }
+
+            foreach (var constraintHandle in parameter.GetConstraints())
+            {
+                EntityHandle constraint =
+                    _reader.GetGenericParameterConstraint(
+                        constraintHandle).Type;
+                if (!ConstraintCanIncludeValueType(constraint))
+                    return false;
+            }
+            return true;
+        }
+        catch (Exception ex) when (ex is BadImageFormatException
+            or InvalidOperationException
+            or ArgumentException
+            or OverflowException
+            or InvalidCastException)
+        {
+            return false;
+        }
+    }
+
+    bool ConstraintCanIncludeValueType(EntityHandle constraint)
+    {
+        if (constraint.Kind == HandleKind.TypeDefinition)
+        {
+            TypeAttributes attributes = _reader
+                .GetTypeDefinition((TypeDefinitionHandle)constraint)
+                .Attributes;
+            return (attributes & TypeAttributes.Interface) != 0;
+        }
+
+        if (constraint.Kind == HandleKind.TypeReference)
+        {
+            var reference = _reader.GetTypeReference(
+                (TypeReferenceHandle)constraint);
+            string @namespace = _reader.GetString(reference.Namespace);
+            string name = _reader.GetString(reference.Name);
+            return @namespace == "System"
+                && name is "ValueType" or "Enum";
+        }
+
+        // Type specifications and generic-parameter constraints cannot be
+        // proven here to admit a value-type instantiation.
+        return false;
+    }
+}
